fix: resolve React fast refresh warning in shared-components
- Extract buildFilterString utility function to separate utils.ts file
- Keep only React components in shared-components.tsx
- Update imports in index.tsx to use new utils file
- Ensures React fast refresh works properly in development
