+++ src/pages/flood-control-projects/utils.ts
@@ -1,0 +1,43 @@
+import { FilterState } from './shared-components';
+
+// Utility function to build filter string
+export const buildFilterString = (filters: FilterState): string => {
+  // Start with an empty array - we'll add filters as needed
+  const filterStrings: string[] = [];
+
+  // Always filter by type
+  filterStrings.push('type = "flood_control"');
+
+  if (filters.InfraYear && filters.InfraYear.trim()) {
+    filterStrings.push(`FundingYear = ${filters.InfraYear.trim()}`);
+  }
+
+  if (filters.Region && filters.Region.trim()) {
+    filterStrings.push(`Region = "${filters.Region.trim()}"`);
+  }
+
+  if (filters.Province && filters.Province.trim()) {
+    filterStrings.push(`Province = "${filters.Province.trim()}"`);
+  }
+
+  if (filters.TypeofWork && filters.TypeofWork.trim()) {
+    filterStrings.push(`TypeofWork = "${filters.TypeofWork.trim()}"`);
+  }
+
+  if (
+    filters.DistrictEngineeringOffice &&
+    filters.DistrictEngineeringOffice.trim()
+  ) {
+    filterStrings.push(
+      `DistrictEngineeringOffice = "${filters.DistrictEngineeringOffice.trim()}"`
+    );
+  }
+
+  if (filters.LegislativeDistrict && filters.LegislativeDistrict.trim()) {
+    filterStrings.push(
+      `LegislativeDistrict = "${filters.LegislativeDistrict.trim()}"`
+    );
+  }
+
+  return filterStrings.length > 0 ? filterStrings.join(' AND ') : '';
+};
